Check FIFOs while waiting for response from kernel driver.
As the fiber link from the MCE operates at 25MB/s and is fed into a pair of
FIFOs that hold 1024 bytes each, the FIFOs fill up in ~80us. Prior to this
change, the firmware (when in c7 mode) did not check the FIFOs from when it
initiated data transfer of a datagram to the host machine, such as for circular
buffer bookkeeping, until after it received confirmation from the host machine
that the datagram had been received and processed by the kernel driver. In the
worst case, when this process started just prior to data being received over
the fiber from the MCE, this meant that the FIFOs would fill up and data would
be lost if this process took more than the ~80us it takes for the FIFOs to
fill up. With this change, the FIFOs are checked while waiting for the kernel
driver, so we are no longer dependent on the kernel driver quickly servicing
its interrupt.

This issue was discovered after a long and painful debugging process into why a
particular hardware setup (Intel Xeon E-2224G CPU installed in an ASRock Rack
C246 WS motherboard) had checksum verification failures several times a minute.
Even when running off identical disk images and with the fiber card installed
in an external PCI-to-PCIe chassis such that there were identical electrical
connections to the fiber card, this computer would experience data corruption
while there would be no issues with the old computer that was being replaced.
For unclear reasons, the problematic computer seems to take longer to initiate
PCI transactions. After this fix was shown to work, the root cause was verified
by soldering a wire to the full flag pin (pin 9) on one of the FIFO chips (U14)
on a fiber card running unmodified U0107 firmware and monitoring it with an
oscilloscope. Sure enough, this showed that the full flag was asserted on order
100us every few seconds and that this was the cause of the data corruption.
While this problematic hardware surfaced the issue that was fixed, it may very
well have caused data corruption on other hardware with less regularity.